requirements/base.txt
Django==5.0.7
django-environ==0.11.2
psycopg2==2.9.9
djangorestframework==3.15.2

whitenoise==6.7.0

django-filter==24.2
django-extensions==3.2.3
django-cors-headers==4.4.0

attrs==23.2.0

djangorestframework-simplejwt==5.3.1

gunicorn==22.0.0

requests==2.32.3
